스크립트 정리 + 공격 상호작용 완료 / 플레이어 방어 재작성 준비
+ WeaponControl.cs로 각 캐릭터별 무기 콜라이더를 껏다 킴
+ 불필요하다고 판단된 스크립트 제거 및 이름 변경
/ 24.02.16

+++ Assets/Scripts/UI/BarGauge/PlayerBar.cs
@@ -24,6 +24,6 @@ public class PlayerBar : BaseGauge
 
     protected override void UpdateUI()
     {
-        currentPlayerHP.fillAmount = player.HP / player.maxhp;
+        currentPlayerHP.fillAmount = player.HP / (float)player.maxhp;
     }
 }
